+++ cypress/support/step-definitions/navigation-bar-steps.js
@@ -1,7 +1,6 @@
 import { Then } from "@badeball/cypress-cucumber-preprocessor";
-
-import navigationBarChildren from "../../locators/navigation-bar";
+import navigationBar from "../../locators/navigation-bar";
 
 Then("Navigation Bar children on preview is set to {word}", (text) => {
-  navigationBarChildren().should("have.text", text);
+  navigationBar().should("have.text", text);
 });