+++ src/components/global-header/index.ts
@@ -1,0 +1,2 @@
+export { default } from "./global-header.component";
+export type { GlobalHeaderProps } from "./global-header.component";